+++ CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-03-22
+
+### Added
+
+- Spinner when converting audio files.
+
+### Fixed
+
+- Proper path to ffmpeg library in conversion utility function.
+- Reload CSS on every change.
+
 ## [0.1.0] - 2022-03-21
 
 ### Changed
 
 - Improve build and dev setup.
 
-[unreleased]: https://github.com/mcjlnrtwcz/tracks-sample-manager/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/mcjlnrtwcz/tracks-sample-manager/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/mcjlnrtwcz/tracks-sample-manager/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/mcjlnrtwcz/tracks-sample-manager/releases/tag/v0.1.0